Add files via upload
Project title: Music Sales Analysis using SQL
Project objective: The objective of this project is to use SQL to analyze music sales data. This will involve performing queries to extract insights into the data, such as the most popular genres of music, the best-selling artists, and the most profitable albums.